deployment-template: Update commands with region
Udpates the `deployment-template` commands to use a `region` on the API
calls to the SDK.
